// src/generated/models/vitals-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Healix Clinic Management API
 * REST API for clinic management system with EHR support for Maharashtra small to mid-sized clinics.  ## Features - Patient registration and management - Emergency contact management - Insurance details tracking - Medical history records - Appointment scheduling and management - Vitals recording and examination - Prescription and investigation management - IPD/OPD workflow 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@healix.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface VitalsRequest {
    /**
     * Weight in kg
     */
    'weight'?: number;
    /**
     * Height in cm
     */
    'height'?: number;
    /**
     * Head circumference in cm (pediatric)
     */
    'headCircumference'?: number;
    'temperature'?: number;
    'temperatureUnit'?: VitalsRequestTemperatureUnitEnum;
    /**
     * Heart rate in bpm
     */
    'heartRate'?: number;
    /**
     * Respiratory rate in breaths/min
     */
    'respiratoryRate'?: number;
    /**
     * Systolic BP in mmHg
     */
    'systolicBp'?: number;
    /**
     * Diastolic BP in mmHg
     */
    'diastolicBp'?: number;
    /**
     * Oxygen saturation %
     */
    'spo2'?: number;
    /**
     * RBS in mg/dL
     */
    'randomBloodSugar'?: number;
    /**
     * Body Mass Index
     */
    'bmi'?: number;
    'bmiStatus'?: VitalsRequestBmiStatusEnum;
    /**
     * Pain level on 0-10 scale
     */
    'painLevel'?: number;
    'symptoms'?: Array<string>;
}

export const VitalsRequestTemperatureUnitEnum = {
    F: 'F',
    C: 'C'
} as const;

export type VitalsRequestTemperatureUnitEnum = typeof VitalsRequestTemperatureUnitEnum[keyof typeof VitalsRequestTemperatureUnitEnum];
export const VitalsRequestBmiStatusEnum = {
    Underweight: 'Underweight',
    Normal: 'Normal',
    Overweight: 'Overweight',
    Obese: 'Obese'
} as const;

export type VitalsRequestBmiStatusEnum = typeof VitalsRequestBmiStatusEnum[keyof typeof VitalsRequestBmiStatusEnum];